Use correct execroot to obtain deployment files.
Currently ASwB assumes a project will only have one execroot.  As a
result the build process will attempt to save time by reusing the
execroot fetched during initial project sync.  This is not always correct
and caused a bug where the build process will use a remote execroot
as it's local blaze execroot and failed due to "missing files".

This CL fixes this by reintroducing per-build execroot fetching.
Fetching the execroot is quick (~1-2 seconds), but that's still an
extra step before every single build.  ExecrootUtil includes a cache
for the default execroot to avoid fetching the default execroot over
and over again.  In the event a user is building with a custom execroot,
they'll still re-fetch their custom execroot.

PiperOrigin-RevId: 287870209